requirements.txt
Flask==2.3.3
SQLAlchemy==2.0.20
Flask-SQLAlchemy==3.1.1
WTForms==3.0.1
Flask-WTF==1.1.1
python-dotenv==1.0.0
